Move pkg/auth to kubeops
