+++ cmd/kubeops/internal/auth/auth_test.go
